#145 Add unit test cases for spark cubing engine
