+++ examples/test_case_data/parquet_test/cube/ci_left_join_cube.json
@@ -1,14 +1,48 @@
 {
   "uuid" : "56e1a32a-a33e-4b69-83dd-0bb8b1f80354",
-  "last_modified" : 1585736800517,
+  "last_modified" : 1585910517131,
   "version" : "3.9.9.1",
   "name" : "ci_left_join_cube",
   "owner" : null,
   "descriptor" : "ci_left_join_cube",
   "display_name" : "ci_left_join_cube",
   "cost" : 50,
-  "status" : "DISABLED",
-  "segments" : [ ],
+  "status" : "READY",
+  "segments" : [ {
+    "uuid" : "6ceb8a86-b2d5-c0ab-d7b8-d1ff99bff859",
+    "name" : "FULL_BUILD",
+    "storage_location_identifier" : "KYLIN_TF0LF0SEAW",
+    "date_range_start" : 0,
+    "date_range_end" : 9223372036854775807,
+    "source_offset_start" : 0,
+    "source_offset_end" : 0,
+    "status" : "READY",
+    "size_kb" : 4528,
+    "is_merged" : false,
+    "estimate_ratio" : null,
+    "input_records" : 10000,
+    "input_records_size" : 2103495,
+    "last_build_time" : 1585910516984,
+    "last_build_job_id" : "91582b8c-6d79-4531-8f9d-3c49e7365814",
+    "create_time_utc" : 1585910494869,
+    "cuboid_shard_nums" : { },
+    "total_shards" : 0,
+    "blackout_cuboids" : [ ],
+    "binary_signature" : null,
+    "dictionaries" : null,
+    "snapshots" : {
+      "DEFAULT.TEST_COUNTRY" : "default/table_snapshot/DEFAULT.TEST_COUNTRY/e137d29b-9231-4305-8c9d-71fcf54bc836",
+      "EDW.TEST_SITES" : "default/table_snapshot/EDW.TEST_SITES/91c1b6bb-e198-4389-9e6e-c5424175c017",
+      "EDW.TEST_SELLER_TYPE_DIM" : "default/table_snapshot/EDW.TEST_SELLER_TYPE_DIM/a21caacb-8b45-43fe-ac8a-63a25ecd5ae1",
+      "EDW.TEST_CAL_DT" : "default/table_snapshot/EDW.TEST_CAL_DT/cd8bfeb7-63bd-4a00-ae03-7a603f1885dc",
+      "DEFAULT.TEST_CATEGORY_GROUPINGS" : "default/table_snapshot/DEFAULT.TEST_CATEGORY_GROUPINGS/7d63670e-4bb8-4cb1-9439-0e4414f952d1"
+    },
+    "rowkey_stats" : [ ],
+    "stream_source_checkpoint" : null,
+    "additionalInfo" : {
+      "storageType" : "4"
+    }
+  } ],
   "create_time_utc" : 0,
   "cuboid_bytes" : null,
   "cuboid_bytes_recommend" : null,
