+++ kylin-spark-project/kylin-spark-engine/src/test/java/org/apache/kylin/engine/spark/job/SparkCubingJobTest.java
@@ -18,28 +18,18 @@
 
 package org.apache.kylin.engine.spark.job;
 
-import java.text.SimpleDateFormat;
-import java.util.HashSet;
-import java.util.Locale;
-import java.util.Map;
-import java.util.Set;
-import java.util.TimeZone;
-
 import org.apache.kylin.common.KylinConfig;
 import org.apache.kylin.cube.CubeInstance;
 import org.apache.kylin.cube.CubeSegment;
 import org.apache.kylin.engine.spark.LocalWithSparkSessionTest;
 import org.apache.kylin.engine.spark.NSparkCubingEngine;
-import org.apache.kylin.engine.spark.builder.CreateFlatTable;
 import org.apache.kylin.engine.spark.metadata.FunctionDesc;
 import org.apache.kylin.engine.spark.metadata.MetadataConverter;
 import org.apache.kylin.engine.spark.metadata.cube.PathManager;
 import org.apache.kylin.engine.spark.metadata.cube.model.LayoutEntity;
-import org.apache.kylin.job.execution.AbstractExecutable;
-import org.apache.kylin.job.execution.CheckpointExecutable;
+import org.apache.kylin.job.exception.SchedulerException;
 import org.apache.kylin.job.execution.ExecutableManager;
 import org.apache.kylin.job.execution.ExecutableState;
-import org.apache.kylin.job.impl.threadpool.DefaultScheduler;
 import org.apache.kylin.metadata.model.IStorageAware;
 import org.apache.kylin.metadata.model.SegmentRange;
 import org.apache.kylin.storage.StorageFactory;
@@ -54,41 +44,33 @@ import org.apache.spark.sql.types.StructField;
 import org.apache.spark.sql.types.StructType;
 import org.apache.spark.sql.udaf.PreciseCountDistinct;
 import org.junit.Assert;
-import org.junit.After;
-import org.junit.Before;
 import org.junit.Test;
 import org.roaringbitmap.longlong.Roaring64NavigableMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.spark_project.guava.collect.Sets;
 
+import java.util.HashSet;
+import java.util.Map;
+import java.util.Set;
+
 public class SparkCubingJobTest extends LocalWithSparkSessionTest {
 
     private static final Logger logger = LoggerFactory.getLogger(SparkCubingJobTest.class);
     private static StructType OUT_SCHEMA = null;
+    private KylinConfig kylinConfig;
 
-    private ExecutableManager jobService;
-
-    @Before
-    public void setup() throws Exception {
-        super.init();
-        final KylinConfig kylinConfig = KylinConfig.getInstanceFromEnv();
+    @Override
+    public void setup() throws SchedulerException {
+        super.setup();
+        kylinConfig = KylinConfig.getInstanceFromEnv();
         kylinConfig.setProperty("kylin.source.provider.0", "org.apache.kylin.engine.spark.source.HiveSource");
-        jobService = ExecutableManager.getInstance(kylinConfig);
-        for (String jobId : jobService.getAllJobIds()) {
-            AbstractExecutable executable = jobService.getJob(jobId);
-            if (executable instanceof CheckpointExecutable) {
-                jobService.deleteJob(jobId);
-            }
-        }
     }
 
-    @After
+    @Override
     public void after() {
-        DefaultScheduler.destroyInstance();
-        super.cleanupTestMetadata();
+        super.after();
     }
-
 
     @Test
     public void testBuildJob() throws Exception {
@@ -96,30 +78,33 @@ public class SparkCubingJobTest extends LocalWithSparkSessionTest {
 
         cleanupSegments(cubeName);
         CubeInstance cubeInstance = cubeMgr.getCube(cubeName);
+        ExecutableManager jobService = ExecutableManager.getInstance(kylinConfig);
 
-        SimpleDateFormat f = new SimpleDateFormat("yyyy-MM-dd", Locale.ROOT);
-        f.setTimeZone(TimeZone.getTimeZone("GMT"));
         long date1 = 0;
-        long date2 = f.parse("2012-06-01").getTime();
-        long date3 = f.parse("2013-07-01").getTime();
+        long date2 = dateToLong("2012-06-01");
+        long date3 = dateToLong("2013-07-01");
 
         CubeSegment segment = cubeMgr.appendSegment(cubeInstance, new SegmentRange.TSRange(date1, date2));
         NSparkCubingJob job = NSparkCubingJob.create(Sets.newHashSet(segment), "ADMIN");
         jobService.addJob(job);
         // wait job done
-        ExecutableState state = waitForJob(job.getId());
+        ExecutableState state = wait(job);
         Assert.assertEquals(ExecutableState.SUCCEED, state);
-
-        CubeSegment segment2 = cubeMgr.appendSegment(cubeInstance, new SegmentRange.TSRange(date2, date3));
-        NSparkCubingJob job2 = NSparkCubingJob.create(Sets.newHashSet(segment2), "ADMIN");
-        jobService.addJob(job2);
-        // wait job done
-        ExecutableState state2 = waitForJob(job2.getId());
-        Assert.assertEquals(ExecutableState.SUCCEED, state2);
 
         // Result cmp: Parquet vs Spark SQL
         queryTest(segment);
         snapshotTest(segment);
+
+        // Test build 2nd segment
+        CubeSegment segment2 = cubeMgr.appendSegment(cubeInstance, new SegmentRange.TSRange(date2, date3));
+        NSparkCubingJob job2 = NSparkCubingJob.create(Sets.newHashSet(segment2), "ADMIN");
+        jobService.addJob(job2);
+        // wait job done
+        ExecutableState state2 = wait(job2);
+        Assert.assertEquals(ExecutableState.SUCCEED, state2);
+
+        cubeInstance = cubeMgr.reloadCube(cubeName);
+        Assert.assertEquals(2, cubeInstance.getSegments().size());
     }
 
     @Test
@@ -127,28 +112,27 @@ public class SparkCubingJobTest extends LocalWithSparkSessionTest {
         String cubeName = "ci_inner_join_cube";
         cleanupSegments(cubeName);
         CubeInstance cubeInstance = cubeMgr.getCube(cubeName);
+        ExecutableManager jobService = ExecutableManager.getInstance(kylinConfig);
 
         /**
          * Round1. Build 2 segment
          */
-        SimpleDateFormat f = new SimpleDateFormat("yyyy-MM-dd", Locale.ROOT);
-        f.setTimeZone(TimeZone.getTimeZone("GMT"));
-        long date1 = f.parse("2010-01-01").getTime();
-        long date2 = f.parse("2013-01-01").getTime();
-        long date3 = f.parse("2014-01-01").getTime();
+        long date1 = dateToLong("2010-01-01");
+        long date2 = dateToLong("2012-01-01");
+        long date3 = dateToLong("2014-01-01");
 
         CubeSegment segment = cubeMgr.appendSegment(cubeInstance, new SegmentRange.TSRange(date1, date2));
         NSparkCubingJob job = NSparkCubingJob.create(Sets.newHashSet(segment), "ADMIN");
         jobService.addJob(job);
         // wait job done
-        ExecutableState state = waitForJob(job.getId());
+        ExecutableState state = wait(job);
         Assert.assertEquals(ExecutableState.SUCCEED, state);
 
         CubeSegment segment2 = cubeMgr.appendSegment(cubeInstance, new SegmentRange.TSRange(date2, date3));
         NSparkCubingJob job2 = NSparkCubingJob.create(Sets.newHashSet(segment2), "ADMIN");
         jobService.addJob(job2);
         // wait job done
-        ExecutableState state2 = waitForJob(job2.getId());
+        ExecutableState state2 = wait(job2);
         Assert.assertEquals(ExecutableState.SUCCEED, state2);
 
         cubeInstance = cubeMgr.reloadCube(cubeName);
@@ -162,15 +146,11 @@ public class SparkCubingJobTest extends LocalWithSparkSessionTest {
         jobService.addJob(firstMergeJob);
         // wait job done
         Assert.assertEquals(ExecutableState.SUCCEED, wait(firstMergeJob));
-    }
 
-    @Test
-    public void testMergeResult() {
-        String cubeName = "ci_inner_join_cube";
-
-        CubeInstance cubeInstance = cubeMgr.getCube(cubeName);
-        for (CubeSegment segment : cubeInstance.getSegments()) {
-            queryTest(segment);
+        // Parquet cuboids vs Spark SQL
+        cubeInstance = cubeMgr.reloadCube(cubeName);
+        for (CubeSegment seg : cubeInstance.getSegments()) {
+            queryTest(seg);
         }
     }
 
@@ -280,29 +260,4 @@ public class SparkCubingJobTest extends LocalWithSparkSessionTest {
 
         return index;
     }
-
-    private Dataset<Row> initFlatTable(CubeSegment segment) {
-        System.out.println(getTestConfig().getMetadataUrl());
-
-        CreateFlatTable flatTable = new CreateFlatTable(
-                MetadataConverter.getSegmentInfo(segment.getCubeInstance(), segment.getUuid(), segment.getName()), null, ss, null);
-        Dataset<Row> ds = flatTable.generateDataset(false, true);
-        return ds;
-    }
-
-    protected ExecutableState waitForJob(String jobId) {
-        while (true) {
-            AbstractExecutable job = jobService.getJob(jobId);
-            if (job.getStatus() == ExecutableState.SUCCEED || job.getStatus() == ExecutableState.ERROR) {
-                return job.getStatus();
-            } else {
-                try {
-                    Thread.sleep(5000);
-                } catch (InterruptedException e) {
-                    e.printStackTrace();
-                }
-            }
-        }
-    }
-
 }
