+++ kylin-spark-project/kylin-spark-test/src/test/java/org/apache/kylin/engine/spark2/NManualBuildAndQueryTest.java
@@ -21,23 +21,18 @@ import com.google.common.collect.Maps;
 import java.util.Map;
 import org.apache.kylin.common.KylinConfig;
 import org.apache.kylin.common.util.Pair;
-import org.apache.kylin.cube.CubeInstance;
 import org.apache.kylin.cube.CubeManager;
 import org.apache.kylin.cube.CubeSegment;
 import org.apache.kylin.engine.spark.LocalWithSparkSessionTest;
-import org.apache.kylin.engine.spark.job.NSparkMergingJob;
-import org.apache.kylin.job.execution.ExecutableManager;
+import org.apache.kylin.job.exception.SchedulerException;
 import org.apache.kylin.job.execution.ExecutableState;
-import org.apache.kylin.job.impl.threadpool.DefaultScheduler;
 import org.apache.kylin.metadata.model.SegmentRange;
 import org.apache.kylin.metadata.realization.RealizationType;
 import org.apache.kylin.query.routing.Candidate;
 import org.apache.spark.sql.KylinSparkEnv;
 import org.apache.kylin.engine.spark2.NExecAndComp.CompareLevel;
 import org.apache.spark.sql.SparderContext;
-import org.junit.After;
 import org.junit.Assert;
-import org.junit.Before;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.slf4j.Logger;
@@ -45,11 +40,8 @@ import org.slf4j.LoggerFactory;
 import org.spark_project.guava.collect.Lists;
 
 import java.io.File;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
-import java.util.TimeZone;
 import java.util.concurrent.Callable;
 import java.util.concurrent.CompletionService;
 import java.util.concurrent.ExecutorCompletionService;
@@ -64,27 +56,21 @@ public class NManualBuildAndQueryTest extends LocalWithSparkSessionTest {
 
     private boolean succeed = true;
 
-    @Before
-    public void setup() throws Exception {
-        super.init();
-        System.setProperty("kylin.env", "UT");
-        System.setProperty("kylin.metadata.distributed-lock-impl", "org.apache.kylin.engine.spark.utils.MockedDistributedLock$MockedFactory");
-        System.setProperty("isDeveloperMode", "true");
-        System.setProperty("kylin.query.enable-dynamic-column", "false");
+    @Override
+    public void setup() throws SchedulerException {
+        super.setup();
+        overwriteSystemProp("kylin.env", "UT");
+        overwriteSystemProp("isDeveloperMode", "true");
+        overwriteSystemProp("kylin.query.enable-dynamic-column", "false");
         Map<RealizationType, Integer> priorities = Maps.newHashMap();
         priorities.put(RealizationType.HYBRID, 0);
         priorities.put(RealizationType.CUBE, 0);
         Candidate.setPriorities(priorities);
-
     }
 
-    @After
+    @Override
     public void after() {
-        DefaultScheduler.destroyInstance();
-        //super.cleanupTestMetadata();
-        System.clearProperty("kylin.job.scheduler.poll-interval-second");
-        System.clearProperty("noBuild");
-        System.clearProperty("isDeveloperMode");
+        super.after();
     }
 
     @Test
@@ -218,7 +204,7 @@ public class NManualBuildAndQueryTest extends LocalWithSparkSessionTest {
 //            fullBuildCube("ci_inner_join_cube");
             fullBuildCube("ci_left_join_cube");
         } else {
-            buildAndMergeCube("ci_inner_join_cube");
+            //buildAndMergeCube("ci_inner_join_cube");
             buildAndMergeCube("ci_left_join_cube");
         }
     }
@@ -235,74 +221,34 @@ public class NManualBuildAndQueryTest extends LocalWithSparkSessionTest {
     private void buildTwoSegementAndMerge(String cubeName) throws Exception {
         KylinConfig config = KylinConfig.getInstanceFromEnv();
         CubeManager cubeMgr = CubeManager.getInstance(config);
-        ExecutableManager execMgr = ExecutableManager.getInstance(config);
         Assert.assertTrue(config.getHdfsWorkingDirectory().startsWith("file:"));
 
         // cleanup all segments first
         cleanupSegments(cubeName);
 
-        SimpleDateFormat f = new SimpleDateFormat("yyyy-MM-dd", Locale.ROOT);
-        f.setTimeZone(TimeZone.getTimeZone("GMT"));
-
         /**
          * Round1. Build 2 segment
          */
-        long start = f.parse("2010-01-01").getTime();
-        long end = f.parse("2013-01-01").getTime();
-        buildCuboid(cubeName, new SegmentRange.TSRange(start, end));
-        start = f.parse("2013-01-01").getTime();
-        end = f.parse("2015-01-01").getTime();
-        buildCuboid(cubeName, new SegmentRange.TSRange(start, end));
+        ExecutableState state;
+        state = buildCuboid(cubeName, new SegmentRange.TSRange(dateToLong("2010-01-01"), dateToLong("2012-01-01")));
+        Assert.assertEquals(ExecutableState.SUCCEED, state);
+
+        state = buildCuboid(cubeName, new SegmentRange.TSRange(dateToLong("2012-01-01"), dateToLong("2015-01-01")));
+        Assert.assertEquals(ExecutableState.SUCCEED, state);
 
         /**
          * Round2. Merge two segments
          */
-
-        CubeInstance cube = cubeMgr.reloadCube(cubeName);
-        CubeSegment firstMergeSeg = cubeMgr.mergeSegments(cube,
-                new SegmentRange.TSRange(f.parse("2010-01-01").getTime(), f.parse("2015-01-01").getTime()), null,
-                false);
-        NSparkMergingJob firstMergeJob = NSparkMergingJob.merge(firstMergeSeg, "ADMIN");
-        execMgr.addJob(firstMergeJob);
-        // wait job done
-        Assert.assertEquals(ExecutableState.SUCCEED, wait(firstMergeJob));
+        state = mergeSegments(cubeName, dateToLong("2010-01-01"), dateToLong("2015-01-01"), false);
+        Assert.assertEquals(ExecutableState.SUCCEED, state);
 
         /**
          * validate cube segment info
          */
         CubeSegment firstSegment = cubeMgr.reloadCube(cubeName).getSegments().get(0);
 
-        /*if (getProject().equals("default") && cubeName.equals("ci_left_join_cube")) {
-            Map<Long, NDataLayout> cuboidsMap1 = firstSegment.();
-            Map<Long, Long[]> compareTuples1 = Maps.newHashMap();
-            compareTuples1.put(1L, new Long[] { 9896L, 9896L });
-            compareTuples1.put(10001L, new Long[] { 9896L, 9896L });
-            compareTuples1.put(10002L, new Long[] { 9896L, 9896L });
-            compareTuples1.put(20001L, new Long[] { 9896L, 9896L });
-            compareTuples1.put(30001L, new Long[] { 9896L, 9896L });
-            compareTuples1.put(1000001L, new Long[] { 9896L, 9896L });
-            compareTuples1.put(1010001L, new Long[] { 731L, 9163L });
-            compareTuples1.put(1020001L, new Long[] { 302L, 6649L });
-            compareTuples1.put(1030001L, new Long[] { 44L, 210L });
-            compareTuples1.put(1040001L, new Long[] { 9163L, 9884L });
-            compareTuples1.put(1050001L, new Long[] { 105L, 276L });
-            compareTuples1.put(1060001L, new Long[] { 138L, 286L });
-            compareTuples1.put(1070001L, new Long[] { 9880L, 9896L });
-            compareTuples1.put(1080001L, new Long[] { 9833L, 9896L });
-            compareTuples1.put(1090001L, new Long[] { 9421L, 9884L });
-            compareTuples1.put(1100001L, new Long[] { 143L, 6649L });
-            compareTuples1.put(1110001L, new Long[] { 4714L, 9884L });
-            compareTuples1.put(1120001L, new Long[] { 9884L, 9896L });
-            compareTuples1.put(20000000001L, new Long[] { 9896L, 9896L });
-            compareTuples1.put(20000010001L, new Long[] { 9896L, 9896L });
-            compareTuples1.put(20000020001L, new Long[] { 9896L, 9896L });
-            verifyCuboidMetrics(cuboidsMap1, compareTuples1);
-        }*/
-
-        Assert.assertEquals(new SegmentRange.TSRange(f.parse("2010-01-01").getTime(), f.parse("2015-01-01").getTime()),
+        Assert.assertEquals(new SegmentRange.TSRange(dateToLong("2010-01-01"), dateToLong("2015-01-01")),
                 firstSegment.getSegRange());
-        //Assert.assertEquals(27, firstSegment.getDictionaries().size());
-        Assert.assertEquals(7, firstSegment.getSnapshots().size());
     }
 
     private void buildFourSegementAndMerge(String cubeName) throws Exception {
@@ -311,47 +257,30 @@ public class NManualBuildAndQueryTest extends LocalWithSparkSessionTest {
         // cleanup all segments first
         cleanupSegments(cubeName);
 
-        SimpleDateFormat f = new SimpleDateFormat("yyyy-MM-dd", Locale.ROOT);
-        f.setTimeZone(TimeZone.getTimeZone("GMT"));
         /**
          * Round1. Build 4 segment
          */
-        long start = f.parse("2010-01-01").getTime();
-        long end = f.parse("2012-06-01").getTime();
-        buildCuboid(cubeName, new SegmentRange.TSRange(start, end));
+        ExecutableState state;
+        state = buildCuboid(cubeName, new SegmentRange.TSRange(dateToLong("2010-01-01"), dateToLong("2012-06-01")));
+        Assert.assertEquals(ExecutableState.SUCCEED, state);
 
-        start = f.parse("2012-06-01").getTime();
-        end = f.parse("2013-01-01").getTime();
-        buildCuboid(cubeName, new SegmentRange.TSRange(start, end));
+        state = buildCuboid(cubeName, new SegmentRange.TSRange(dateToLong("2012-06-01"), dateToLong("2013-01-01")));
+        Assert.assertEquals(ExecutableState.SUCCEED, state);
 
-        start = f.parse("2013-01-01").getTime();
-        end = f.parse("2013-06-01").getTime();
-        buildCuboid(cubeName, new SegmentRange.TSRange(start, end));
+        state = buildCuboid(cubeName, new SegmentRange.TSRange(dateToLong("2013-01-01"), dateToLong("2013-06-01")));
+        Assert.assertEquals(ExecutableState.SUCCEED, state);
 
-        start = f.parse("2013-06-01").getTime();
-        end = f.parse("2015-01-01").getTime();
-        buildCuboid(cubeName, new SegmentRange.TSRange(start, end));
+        state = buildCuboid(cubeName, new SegmentRange.TSRange(dateToLong("2013-06-01"), dateToLong("2015-01-01")));
+        Assert.assertEquals(ExecutableState.SUCCEED, state);
 
         /**
          * Round2. Merge two segments
          */
-        CubeInstance cube = cubeMgr.reloadCube(cubeName);
-        CubeSegment firstMergeSeg = cubeMgr.mergeSegments(cube,
-                new SegmentRange.TSRange(f.parse("2010-01-01").getTime(), f.parse("2013-01-01").getTime()), null,
-                false);
-        NSparkMergingJob firstMergeJob = NSparkMergingJob.merge(firstMergeSeg, "ADMIN");
-        execMgr.addJob(firstMergeJob);
-        // wait job done
-        Assert.assertEquals(ExecutableState.SUCCEED, wait(firstMergeJob));
+        state = mergeSegments(cubeName, dateToLong("2010-01-01"), dateToLong("2013-01-01"), false);
+        Assert.assertEquals(ExecutableState.SUCCEED, state);
 
-        cube = cubeMgr.reloadCube(cubeName);
-
-        CubeSegment secondMergeSeg = cubeMgr.mergeSegments(cube, new SegmentRange.TSRange(
-                f.parse("2013-01-01").getTime(), f.parse("2015-01-01").getTime()), null, false);
-        NSparkMergingJob secondMergeJob = NSparkMergingJob.merge(secondMergeSeg, "ADMIN");
-        execMgr.addJob(secondMergeJob);
-        // wait job done
-        Assert.assertEquals(ExecutableState.SUCCEED, wait(secondMergeJob));
+        state = mergeSegments(cubeName, dateToLong("2013-01-01"), dateToLong("2015-01-01"), false);
+        Assert.assertEquals(ExecutableState.SUCCEED, state);
 
         /**
          * validate cube segment info
@@ -359,40 +288,10 @@ public class NManualBuildAndQueryTest extends LocalWithSparkSessionTest {
         CubeSegment firstSegment = cubeMgr.reloadCube(cubeName).getSegments().get(0);
         CubeSegment secondSegment = cubeMgr.reloadCube(cubeName).getSegments().get(1);
 
-        /*if (getProject().equals("default") && cubeName.equals("ci_inner_join_cube")) {
-            Map<Long, NDataLayout> cuboidsMap1 = firstSegment.getLayoutsMap();
-            Map<Long, Long[]> compareTuples1 = Maps.newHashMap();
-            compareTuples1.put(1L, new Long[] { 4903L, 4903L });
-            compareTuples1.put(10001L, new Long[] { 4903L, 4903L });
-            compareTuples1.put(10002L, new Long[] { 4903L, 4903L });
-            compareTuples1.put(20001L, new Long[] { 4903L, 4903L });
-            compareTuples1.put(30001L, new Long[] { 4903L, 4903L });
-            compareTuples1.put(1000001L, new Long[] { 4903L, 4903L });
-            compareTuples1.put(20000000001L, new Long[] { 4903L, 4903L });
-            compareTuples1.put(20000010001L, new Long[] { 4903L, 4903L });
-            compareTuples1.put(20000020001L, new Long[] { 4903L, 4903L });
-            verifyCuboidMetrics(cuboidsMap1, compareTuples1);
-
-            Map<Long, NDataLayout> cuboidsMap2 = secondSegment.getLayoutsMap();
-            Map<Long, Long[]> compareTuples2 = Maps.newHashMap();
-            compareTuples2.put(1L, new Long[] { 5097L, 5097L });
-            compareTuples2.put(10001L, new Long[] { 5097L, 5097L });
-            compareTuples2.put(10002L, new Long[] { 5097L, 5097L });
-            compareTuples2.put(20001L, new Long[] { 5097L, 5097L });
-            compareTuples2.put(30001L, new Long[] { 5097L, 5097L });
-            compareTuples2.put(1000001L, new Long[] { 5097L, 5097L });
-            compareTuples2.put(20000000001L, new Long[] { 5097L, 5097L });
-            compareTuples2.put(20000010001L, new Long[] { 5097L, 5097L });
-            compareTuples2.put(20000020001L, new Long[] { 5097L, 5097L });
-            verifyCuboidMetrics(cuboidsMap2, compareTuples2);
-        }*/
-
-        Assert.assertEquals(new SegmentRange.TSRange(f.parse("2010-01-01").getTime(), f.parse("2013-01-01").getTime()),
+        Assert.assertEquals(new SegmentRange.TSRange(dateToLong("2010-01-01"), dateToLong("2013-01-01")),
                 firstSegment.getSegRange());
-        Assert.assertEquals(new SegmentRange.TSRange(f.parse("2013-01-01").getTime(), f.parse("2015-01-01").getTime()),
+        Assert.assertEquals(new SegmentRange.TSRange(dateToLong("2013-01-01"), dateToLong("2015-01-01")),
                 secondSegment.getSegRange());
-        //Assert.assertEquals(31, firstSegment.getDictionaries().size());
-        //Assert.assertEquals(31, secondSegment.getDictionaries().size());
     }
 
     class QueryCallable implements Callable<Pair<String, Throwable>> {
