--- spark-project/engine-spark/src/test/scala/io/kyligence/kap/engine/spark/builder/MockDFSnapshotBuilder.scala
@@ -1,44 +1,0 @@
-/*
- * Copyright (C) 2016 Kyligence Inc. All rights reserved.
- * http://kyligence.io
- * This software is the confidential and proprietary information of
- * Kyligence Inc. ("Confidential Information"). You shall not disclose
- * such Confidential Information and shall use it only in accordance
- * with the terms of the license agreement you entered into with
- * Kyligence Inc.
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- */
-package io.kyligence.kap.engine.spark.builder
-
-import io.kyligence.kap.metadata.cube.model.NDataSegment
-import org.apache.kylin.metadata.model.TableDesc
-import org.apache.spark.sql.{Dataset, Row, SparkSession}
-
-class MockDFSnapshotBuilder extends DFSnapshotBuilder {
-
-  var ds: Dataset[Row] = _
-
-  def this(seg: NDataSegment, ss: SparkSession) {
-    this()
-    this.seg = seg
-    this.ss = ss
-  }
-
-  override def getSourceData(tableDesc: TableDesc): Dataset[Row] = {
-    if (ds == null) {
-      ds = super.getSourceData(tableDesc).distinct()
-    }
-    ds
-  }
-}
