--- spark-project/spark-common/src/test/java/org/apache/spark/sql/common/SparderBaseFunSuite.scala
@@ -1,86 +1,0 @@
-/*
- * Copyright (C) 2016 Kyligence Inc. All rights reserved.
- * http://kyligence.io
- * This software is the confidential and proprietary information of
- * Kyligence Inc. ("Confidential Information"). You shall not disclose
- * such Confidential Information and shall use it only in accordance
- * with the terms of the license agreement you entered into with
- * Kyligence Inc.
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
- * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
- * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
- * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
- * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
- * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
- * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- *
- */
-package org.apache.spark.sql.common
-
-// scalastyle:off
-import java.io.File
-
-import org.apache.spark.internal.Logging
-import org.apache.spark.util.{AccumulatorContext, Utils}
-import org.scalatest.{BeforeAndAfterAll, FunSuite, Outcome}
-
-trait SparderBaseFunSuite
-  extends FunSuite
-    with BeforeAndAfterAll
-    with Logging {
-  // scalastyle:on
-
-  protected override def afterAll(): Unit = {
-    try {
-      // Avoid leaking map entries in tests that use accumulators without SparkContext
-      AccumulatorContext.clear()
-    } finally {
-      super.afterAll()
-    }
-  }
-
-  // helper function
-  protected final def getTestResourceFile(file: String): File = {
-    new File(getClass.getClassLoader.getResource(file).getFile)
-  }
-
-  protected final def getTestResourcePath(file: String): String = {
-    getTestResourceFile(file).getCanonicalPath
-  }
-
-  /**
-    * Generates a temporary path without creating the actual file/directory, then pass it to `f`. If
-    * a file/directory is created there by `f`, it will be delete after `f` returns.
-    *
-    * @todo Probably this method should be moved to a more general place
-    */
-  protected def withTempPath(f: File => Unit): Unit = {
-    val path = Utils.createTempDir()
-    path.delete()
-    try f(path) finally Utils.deleteRecursively(path)
-  }
-
-  /**
-    * Log the suite name and the test name before and after each test.
-    *
-    * Subclasses should never override this method. If they wish to run
-    * custom code before and after each test, they should mix in the
-    * {{org.scalatest.BeforeAndAfter}} trait instead.
-    */
-  final protected override def withFixture(test: NoArgTest): Outcome = {
-    val testName = test.text
-    val suiteName = this.getClass.getName
-    val shortSuiteName = suiteName.replaceAll("org.apache.spark", "o.a.s")
-    try {
-      logInfo(s"\n\n===== TEST OUTPUT FOR $shortSuiteName: '$testName' =====\n")
-      test()
-    } finally {
-      logInfo(s"\n\n===== FINISHED $shortSuiteName: '$testName' =====\n")
-    }
-  }
-
-}
